スロットの種類を追加する準備
- パッケージ分け
- クラスの抽象化
- リネーム
- パラメータ追加
