+++ app/src/main/AndroidManifest.xml
@@ -10,15 +10,15 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".activity.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".OcrActivity" />
-        <activity android:name=".SlotActivity" />
+        <activity android:name=".activity.OcrActivity" />
+        <activity android:name=".activity.SlotActivity" />
     </application>
 
 </manifest>
