+++ app/src/main/res/layout/slot_item_name.xml
